+++ src/main/scala/ecs/Entity.scala
@@ -1,63 +1,69 @@
 package ecs
 
-import eventSystem.EventListener
 import identifier.{ID, Identifier}
 
-import scala.collection.mutable
-
-class Entity(val name: String = "") extends EventListener with Identifier {
+class Entity(val name: String = "") extends Identifier {
   override val identifier: ID = if (name.nonEmpty) ID(name) else ID()
-  private val components: mutable.Set[ID] = mutable.Set.empty
-  ECS.addEntity(this)
 
-  def deleteSelf(): Unit = {
-    ECS.removeEntity(this)
-  }
+  private val entityManager = EntityManager
+  // componentType, (componentUuid, componentIndex)
+  var componentMap = Map.empty[Int, Map[String, Int]]
+  entityManager.addEntity(id, this)
 
-  def addComponent[A <: Component](component: A)(implicit componentId: ComponentId[A]): Entity = {
-    components += componentId.identifier
-    component.init(identifier)
-    ECS.addComponent(this, component)
+  def addComponent[A <: Component](
+      component: A
+  )(implicit componentId: ComponentId[A]): this.type = {
+    component.entityId = id
+    val componentType = componentId.id
+    val pos = entityManager.addComponent(component)
+    val componentList = componentMap.getOrElse(componentType, Map.empty)
+    val newComponent = component.uuid -> pos
+    componentMap += (componentType -> (componentList + newComponent))
     this
   }
 
-  def removeComponent[A <: Component](implicit component: ComponentId[A]): Entity = {
-    components -= component.identifier
-    ECS.removeComponent(this, component)
+  def getSiblings[A <: Component](componentId: Int): Option[Seq[A]] = {
+    for {
+      componentMap <- componentMap.get(componentId)
+      sibling <- entityManager.getComponents[A](componentId, componentMap.values.toList)
+    } yield sibling
+  }
+
+  def getSibling[A <: Component](componentId: Int): Option[A] = {
+    for {
+      componentMap <- componentMap.get(componentId)
+      componentArrayIdx = componentMap.values.head
+      sibling <- entityManager.getComponents[A](componentId, Seq(componentArrayIdx))
+    } yield sibling.head
+  }
+
+  def demandSiblings[A <: Component](componentId: Int): Seq[A] =
+    getSiblings[A](componentId).getOrElse(
+      throw new RuntimeException(s"could not get component $componentId")
+    )
+
+  def demandSibling[A <: Component](componentId: Int): A = demandSiblings[A](componentId).head
+
+  def removeComponent[A <: Component](
+      componentUuid: Option[String] = None
+  )(implicit componentId: ComponentId[A]): this.type = {
+    val componentType = componentId.id
+    val componentArrayIdx = componentMap.get(componentType) match {
+      case None => None
+      case Some(components) =>
+        componentUuid match {
+          case Some(uuid) =>
+            val componentArrayIdx = components.get(uuid)
+            val componentList = componentMap.getOrElse(componentType, Map.empty)
+            componentMap += (componentType -> (componentList - uuid))
+            componentArrayIdx
+          case None =>
+            val componentArrayIdx = components.headOption.map(_._2)
+            componentMap -= componentType
+            componentArrayIdx
+        }
+    }
+    for { idx <- componentArrayIdx } entityManager.removeComponent(componentType, idx)
     this
   }
-
-  def getComponent[A <: Component](implicit componentId: ComponentId[A]): Option[List[A]] = {
-    if (this.hasComponent[A]) ECS.getEntityComponent(this)
-    else None
-  }
-
-  def hasComponent[A <: Component](implicit component: ComponentId[A]): Boolean =
-    components.contains(component.identifier)
-
-  def hasComponents[A <: Component, B <: Component](implicit
-      component: ComponentId[A],
-      component2: ComponentId[B]
-  ): Boolean =
-    List(component, component2).map(_.identifier).toSet.subsetOf(components)
-
-  def hasComponents[A <: Component, B <: Component, C <: Component](implicit
-      component: ComponentId[A],
-      component2: ComponentId[B],
-      component3: ComponentId[C]
-  ): Boolean =
-    List(component, component2, component3).map(_.identifier).toSet.subsetOf(components)
-
-  def hasComponents[A <: Component, B <: Component, C <: Component, D <: Component](implicit
-      component: ComponentId[A],
-      component2: ComponentId[B],
-      component3: ComponentId[C],
-      component4: ComponentId[D]
-  ): Boolean =
-    List(
-      component,
-      component2,
-      component3,
-      component4
-    ).map(_.identifier).toSet.subsetOf(components)
 }
