+++ src/main/scala/ecs/SingletonComponent.scala
@@ -1,0 +1,5 @@
+package ecs
+
+trait SingletonComponent extends Component {
+  override val singleton = true
+}
