+++ src/main/scala/eventSystem/Events.scala
@@ -5,7 +5,6 @@ import identifier.Identifier
 import org.joml.Matrix4f
 
 object EventId {
-  // TODO could a macro save some ball ache?
   implicit val gameLoopStart: EventId[GameLoopStart] = new EventId
   implicit val gameLoopTick: EventId[GameLoopTick] = new EventId
   implicit val gameLoopEnd: EventId[GameLoopEnd] = new EventId
@@ -21,21 +20,12 @@ object EventId {
   implicit val cameraMove: EventId[CameraMove] = new EventId
   implicit val entityCreated: EventId[EntityCreated] = new EventId
   implicit val entityDeleted: EventId[EntityDestroyed] = new EventId
-
-  implicit val componentTransformCreated: EventId[ComponentCreatedTransform] = new EventId
-  implicit val componentModelCreated: EventId[ComponentCreatedModel] = new EventId
-  implicit val componentCameraCreated: EventId[ComponentCreatedCamera] = new EventId
-  implicit val componentLightCreated: EventId[ComponentCreatedLight] = new EventId
-  implicit val componentPlayerMovementCreated: EventId[ComponentCreatedPlayerMovement] = new EventId
-  implicit val componentCreatedCollider: EventId[ComponentCreatedCollider] = new EventId
-  implicit val componentCreatedRigidBody: EventId[ComponentCreatedRigidBody] = new EventId
 }
 
 class EventId[E] extends Identifier
 
 sealed trait Event
 
-// TODO: make all these pass a data val and use in callback to aid extraction
 // Game Events -------------------------------------------------------
 final case class GameLoopStart() extends Event
 final case class GameLoopTick(timeElapsed: Float) extends Event
@@ -57,16 +47,6 @@ final case class CameraMove(transform: Matrix4f) extends Event
 // Entity events ------------------------------------------------------------
 final case class EntityCreated(entity: Entity) extends Event
 final case class EntityDestroyed(entity: Entity) extends Event
-
-// Component events --------------------------------------------------
-final case class ComponentCreatedTransform(component: Transform, entity: Entity) extends Event
-final case class ComponentCreatedModel(component: Model, entity: Entity) extends Event
-final case class ComponentCreatedCamera(component: Camera, entity: Entity) extends Event
-final case class ComponentCreatedLight(component: Light, entity: Entity) extends Event
-final case class ComponentCreatedPlayerMovement(component: PlayerMovement, entity: Entity)
-    extends Event
-final case class ComponentCreatedCollider(component: Collider, entity: Entity) extends Event
-final case class ComponentCreatedRigidBody(component: RigidBody, entity: Entity) extends Event
 
 // Debug/Random Events -----------------------------------------------
 final case class EmptyEvent() extends Event
