+++ src/main/scala/identifier/Identifier.scala
@@ -4,5 +4,5 @@ import identifier.ID.IdValue
 
 trait Identifier {
   val identifier: ID = ID()
-  val id: IdValue = identifier.id
+  def id: IdValue = identifier.id
 }
