+++ src/main/scala/identifier/Uuid.scala
@@ -1,0 +1,7 @@
+package identifier
+
+import java.util.UUID
+
+trait Uuid {
+  lazy val uuid: String = UUID.randomUUID().toString
+}
