--- src/main/scala/input/K.scala
@@ -1,8 +1,0 @@
-package input
-
-import org.lwjgl.glfw.GLFW._
-
-class K {
-  val a: Int = GLFW_KEY_A
-  val b: Int = GLFW_KEY_B
-}
